+++ examples/redis_client.cpp
@@ -24,8 +24,23 @@
 
 #include <iostream>
 
+#ifdef _WIN32
+#include <Winsock2.h>
+#endif /* _WIN32 */
+
 int
 main(void) {
+#ifdef _WIN32
+  //! Windows netword DLL init
+  WORD version = MAKEWORD(2, 2);
+  WSADATA data;
+
+  if (WSAStartup(version, &data) != 0) {
+    std::cerr << "WSAStartup() failure" << std::endl;
+    return -1;
+  }
+#endif /* _WIN32 */
+
   //! Enable logging
   cpp_redis::active_logger = std::unique_ptr<cpp_redis::logger>(new cpp_redis::logger);
 
@@ -62,8 +77,12 @@ main(void) {
   // synchronous commit, no timeout
   client.sync_commit();
 
-  // synchronous commit, timeout
-  // client.sync_commit(std::chrono::milliseconds(100));
+// synchronous commit, timeout
+// client.sync_commit(std::chrono::milliseconds(100));
+
+#ifdef _WIN32
+  WSACleanup();
+#endif /* _WIN32 */
 
   return 0;
 }