+++ includes/cpp_redis/network/redis_connection.hpp
@@ -28,7 +28,12 @@
 #include <vector>
 
 #include <cpp_redis/builders/reply_builder.hpp>
-#include <cpp_redis/network/tcp_client.hpp>
+
+#include <tacopie/tacopie>
+
+#ifndef __CPP_REDIS_READ_SIZE
+#define __CPP_REDIS_READ_SIZE 4096
+#endif /* __CPP_REDIS_READ_SIZE */
 
 namespace cpp_redis {
 
@@ -37,7 +42,7 @@ namespace network {
 class redis_connection {
 public:
   //! ctor & dtor
-  redis_connection(const std::shared_ptr<io_service>& IO);
+  redis_connection(void);
   ~redis_connection(void);
 
   //! copy ctor & assignment operator
@@ -51,7 +56,7 @@ public:
   void connect(const std::string& host = "127.0.0.1", std::size_t port = 6379,
     const disconnection_handler_t& disconnection_handler = nullptr,
     const reply_callback_t& reply_callback               = nullptr);
-  void disconnect(void);
+  void disconnect(bool wait_for_removal = false);
   bool is_connected(void);
 
   //! send cmd
@@ -62,14 +67,17 @@ public:
 
 private:
   //! receive & disconnection handlers
-  bool tcp_client_receive_handler(network::tcp_client&, const std::vector<char>& buffer);
-  void tcp_client_disconnection_handler(network::tcp_client&);
+  void tcp_client_receive_handler(const tacopie::tcp_client::read_result& result);
+  void tcp_client_disconnection_handler(void);
 
   std::string build_command(const std::vector<std::string>& redis_cmd);
 
 private:
+  void call_disconnection_handler(void);
+
+private:
   //! tcp client for redis connection
-  network::tcp_client m_client;
+  tacopie::tcp_client m_client;
 
   //! reply callback
   reply_callback_t m_reply_callback;