+++ includes/cpp_redis/redis_subscriber.hpp
@@ -34,7 +34,7 @@ namespace cpp_redis {
 class redis_subscriber {
 public:
   //! ctor & dtor
-  redis_subscriber(const std::shared_ptr<network::io_service>& IO = nullptr);
+  redis_subscriber(void);
   ~redis_subscriber(void);
 
   //! copy ctor & assignment operator